feat(): add position details deeplink
