+++ packages/widget/src/hooks/use-handle-deep-links.ts
@@ -1,9 +1,12 @@
+import { useInitQueryParams } from "@sk-widget/hooks/use-init-query-params";
 import { useUpdateEffect } from "@sk-widget/hooks/use-update-effect";
 import { usePendingActionDeepLink } from "@sk-widget/pages/details/earn-page/state/use-pending-action-deep-link";
+import { useMountAnimation } from "@sk-widget/providers/mount-animation";
 import {
   usePendingActionDispatch,
   usePendingActionState,
 } from "@sk-widget/providers/pending-action-state";
+import { useSKWallet } from "@sk-widget/providers/sk-wallet";
 import { Maybe } from "purify-ts";
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
@@ -14,22 +17,46 @@ export const useHandleDeepLinks = () => {
   const navigateRef = useSavedRef(useNavigate());
   const pendingActionState = usePendingActionState();
   const pendignActionRequestDispatch = usePendingActionDispatch();
+  const initQueryParams = useInitQueryParams();
+
+  const { mountAnimationFinished } = useMountAnimation();
+  const { isConnected } = useSKWallet();
+
+  const appReady = mountAnimationFinished && isConnected;
+
+  // Position details page
+  useEffect(() => {
+    initQueryParams
+      .filter((val) =>
+        Boolean(val.yieldId && val.balanceId && !val.pendingaction && appReady)
+      )
+      .ifJust((val) =>
+        navigateRef.current(`positions/${val.yieldId}/${val.balanceId}`)
+      );
+  }, [initQueryParams, navigateRef, appReady]);
 
   // Select validators on position details page
   useEffect(() => {
-    Maybe.fromNullable(pendingActionDeepLinkCheck.data).ifJust((val) => {
-      if (val.type === "positionDetails") {
+    Maybe.fromNullable(pendingActionDeepLinkCheck.data)
+      .filter(
+        (val): val is Extract<typeof val, { type: "positionDetails" }> =>
+          appReady && val.type === "positionDetails"
+      )
+      .ifJust((val) =>
         navigateRef.current(
           `positions/${val.yieldOp.id}/${val.balanceId}/select-validator/${val.pendingAction.type}`
-        );
-      }
-    });
-  }, [navigateRef, pendingActionDeepLinkCheck.data]);
+        )
+      );
+  }, [navigateRef, pendingActionDeepLinkCheck.data, appReady]);
 
   // Review pending action
   useEffect(() => {
-    Maybe.fromNullable(pendingActionDeepLinkCheck.data).ifJust((val) => {
-      if (val.type === "review") {
+    Maybe.fromNullable(pendingActionDeepLinkCheck.data)
+      .filter(
+        (val): val is Extract<typeof val, { type: "review" }> =>
+          appReady && val.type === "review"
+      )
+      .ifJust((val) =>
         pendignActionRequestDispatch(
           Maybe.of({
             actionDto: Maybe.empty(),
@@ -43,21 +70,18 @@ export const useHandleDeepLinks = () => {
             interactedToken: val.balance.token,
             pendingActionType: val.pendingActionDto.requestDto.type,
           })
-        );
-      }
-    });
-  }, [pendignActionRequestDispatch, pendingActionDeepLinkCheck.data]);
+        )
+      );
+  }, [pendignActionRequestDispatch, pendingActionDeepLinkCheck.data, appReady]);
 
   useUpdateEffect(() => {
-    pendingActionState.ifJust(() => {
-      if (
-        pendingActionDeepLinkCheck.data?.type === "review" &&
-        pendingActionState.isJust()
-      ) {
+    pendingActionState
+      .chainNullable(() => pendingActionDeepLinkCheck.data)
+      .filter(() => appReady)
+      .ifJust((val) =>
         navigateRef.current(
-          `positions/${pendingActionDeepLinkCheck.data.yieldOp.id}/${pendingActionDeepLinkCheck.data.balanceId}/pending-action/review`
-        );
-      }
-    });
-  }, [pendingActionState, pendingActionDeepLinkCheck.data]);
+          `positions/${val.yieldOp.id}/${val.balanceId}/pending-action/review`
+        )
+      );
+  }, [pendingActionState, pendingActionDeepLinkCheck.data, appReady]);
 };
