+++ packages/widget/src/hooks/use-init-query-params.ts
@@ -1,0 +1,125 @@
+import type { TokenString } from "@sk-widget/domain/types";
+import {
+  type SupportedSKChains,
+  isSupportedChain,
+} from "@sk-widget/domain/types/chains";
+import { useSettings } from "@sk-widget/providers/settings";
+import { MaybeWindow } from "@sk-widget/utils/maybe-window";
+import { ActionTypes } from "@stakekit/api-hooks";
+import { Codec, Left, Right, string } from "purify-ts";
+import { useMemo } from "react";
+
+export const useInitQueryParams = () => {
+  const { externalProviders } = useSettings();
+
+  return useMemo(
+    () =>
+      getAndValidateInitParams({
+        externalProviderInitToken: externalProviders?.initToken,
+      }),
+    [externalProviders?.initToken]
+  );
+};
+
+const pendingActionCodec = Codec.custom<ActionTypes>({
+  decode: (val) =>
+    string
+      .decode(val)
+      .chain((v) =>
+        v in ActionTypes
+          ? Right(v as ActionTypes)
+          : Left("invalid pending action")
+      ),
+  encode: (val) => val,
+});
+
+const skSupportedChainsCodec = Codec.custom<SupportedSKChains>({
+  decode: (val) =>
+    string
+      .decode(val)
+      .chain((v) => (isSupportedChain(v) ? Right(v) : Left("invalid chain"))),
+  encode: (val) => val,
+});
+
+const pathTraversalRegEx = /(\.\.\/)|(\.\/)/;
+
+const safeParamCodec = Codec.custom<string>({
+  decode: (val) =>
+    string
+      .decode(val)
+      .chain((v) =>
+        pathTraversalRegEx.test(v) ? Left("invalid string value") : Right(v)
+      ),
+  encode: (val) => val,
+});
+
+const yieldIdCodec = Codec.custom<string>({
+  decode: (val) =>
+    string.decode(val).chain((v) => {
+      const [network, token, ...yieldName] = v.split("-");
+
+      if (!network || !token || !yieldName.toString()) {
+        return Left("invalid yieldId format");
+      }
+
+      return Right(v);
+    }),
+  encode: (val) => val,
+});
+
+const tabCodec = Codec.custom<"earn" | "positions">({
+  decode: (val) =>
+    val === "earn" || val === "positions" ? Right(val) : Left("invalid chain"),
+  encode: (val) => val,
+});
+
+const accountIdCodec = Codec.custom<string>({
+  decode: (val) => string.decode(val).map((v) => decodeURIComponent(v)),
+  encode: (val) => val,
+});
+
+export const getAndValidateInitParams = ({
+  externalProviderInitToken,
+}: { externalProviderInitToken?: TokenString }) =>
+  MaybeWindow.map((w) => new URL(w.location.href)).map((url) => ({
+    network: safeParamCodec
+      .decode(url.searchParams.get("network"))
+      .chain(skSupportedChainsCodec.decode)
+      .toMaybe()
+      .extractNullable(),
+    token: safeParamCodec
+      .decode(url.searchParams.get("token") ?? externalProviderInitToken)
+      .toMaybe()
+      .extractNullable(),
+    yieldId: safeParamCodec
+      .decode(url.searchParams.get("yieldId"))
+      .chain(yieldIdCodec.decode)
+      .toMaybe()
+      .extractNullable(),
+    balanceId: safeParamCodec
+      .decode(url.searchParams.get("balanceId"))
+      .toMaybe()
+      .extractNullable(),
+    validator: string // Not safeParamCodec as it maybe has ../ or ./
+      .decode(url.searchParams.get("validator"))
+      .toMaybe()
+      .extractNullable(),
+    pendingaction: safeParamCodec
+      .decode(url.searchParams.get("pendingaction"))
+      .chain(pendingActionCodec.decode)
+      .toMaybe()
+      .extractNullable(),
+    referralCode: safeParamCodec
+      .decode(url.searchParams.get("ref"))
+      .toMaybe()
+      .extractNullable(),
+    accountId: accountIdCodec // Not safeParamCodec as it maybe has ../ or ./
+      .decode(url.searchParams.get("accountId"))
+      .toMaybe()
+      .extractNullable(),
+    tab: safeParamCodec
+      .decode(url.searchParams.get("tab"))
+      .chain(tabCodec.decode)
+      .toMaybe()
+      .extractNullable(),
+  }));
